requirements.txt
pandas
numpy
streamlit
scikit-learn
altair
vega_datasets
plotly.express